Cmake Builder: Filter defines for posix paths
CMake requires that posix paths be provided on the command line.
This presents an issue on Windows as Windows like paths can easily slip
into a CMake define through attributes like
`.libs` or `.directories` or `.prefix`.
Handling this from the package recipes would place a burden on all users
to have to consider this Windows oddity.
Instead handle this from the CMake builder class.
If a definition is a valid path, we format it as a posix path.
Additionally, we now inform CMake that such attributes are in fact paths
instead of the current behavior of defaulting to strings.