relocate win rpath: ensure wrapper is aware of install root
To determine when relocation is unneccesary post install (i.e. for another spack dep) we need to know what defines a "spack installed dep".
We do this by informing the wrapper of the spack install prefix root so it can determine if any of the "rpath"'d dependencies have the same prefix.